Use same function for checking refresh at both points
